feat(#529): connect config
Connect-config-parameter:
- debug-log-flag
- dataset-upload directory
- checkpoint directory
- token-key-path
- token-expire-time
- Sqlite-database is not persistent and use the path
  from the config-file.

+++ src/hanami/src/api/http_endpoints/auth/auth_structs.rs
@@ -20,7 +20,7 @@ use schemars::JsonSchema;
 pub struct UserTokenResp {
     pub access_token: String,
     pub token_type: String,
-    pub expires: u32,
+    pub expires: u64,
 }
 
 #[derive(Debug, Deserialize)]
